rename: project fix: Class naming pattern
