+++ logic_layer/context.py
@@ -1,0 +1,10 @@
+__all__ = ("BaseProcessorContext",)
+
+import dataclasses
+
+from . import processors
+
+
+@dataclasses.dataclass
+class BaseProcessorContext:
+    process: "processors.BaseProcessor"
