+++ logic_layer/results.py
@@ -3,7 +3,7 @@ import typing
 
 
 @dataclasses.dataclass
-class ProcessResult:
+class ProcessorResult:
     pre_run: typing.Dict[typing.Callable, typing.Any] = dataclasses.field(default_factory=dict)
     run: typing.Any = None
     post_run: typing.Dict[typing.Callable, typing.Any] = dataclasses.field(default_factory=dict)
