--- logic_processors/context.py
@@ -1,10 +1,0 @@
-__all__ = ("BaseProcessContext",)
-
-import dataclasses
-
-from . import processors
-
-
-@dataclasses.dataclass
-class BaseProcessContext:
-    process: "processors.BaseProcess"
