+++ setup.py
@@ -1,9 +1,9 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="logic_processors",
-    version="1",
-    description="Mini framework for create logic",
+    name="logic_layer",
+    version="1.1",
+    description="Abstractions for create business logic",
     url="https://github.com/GefMar/logic_processors",
     author="Sergei (Gefest) Romanchuk",
     license="MIT",
